removed is within note and beefed up is-selection-witihin-note
